requirements.txt
aiocache==0.12.1 ; python_version >= "3.11" and python_version < "4.0"
attrs==23.1.0 ; python_version >= "3.11" and python_version < "4.0"
numpy==1.25.0 ; python_version >= "3.11" and python_version < "4.0"
pandas==2.0.3 ; python_version >= "3.11" and python_version < "4.0"
pillow==9.5.0 ; python_version >= "3.11" and python_version < "4.0"
pyqt6-qt6==6.5.1 ; python_version >= "3.11" and python_version < "4.0"
pyqt6-sip==13.5.1 ; python_version >= "3.11" and python_version < "4.0"
pyqt6==6.5.1 ; python_version >= "3.11" and python_version < "4.0"
pyqtconfig==0.9.2 ; python_version >= "3.11" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.11" and python_version < "4.0"
python-mpd2==3.1.0 ; python_version >= "3.11" and python_version < "4.0"
pytz==2023.3 ; python_version >= "3.11" and python_version < "4.0"
pyyaml==6.0 ; python_version >= "3.11" and python_version < "4.0"
qtinter==0.11.0 ; python_version >= "3.11" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.11" and python_version < "4.0"
tzdata==2023.3 ; python_version >= "3.11" and python_version < "4.0"
xdg==6.0.0 ; python_version >= "3.11" and python_version < "4.0"
